requirements.txt
Django==5.2.8
djangorestframework
gunicorn
whitenoise
dj-database-url
psycopg2-binary
